backend/requirements.txt
# Core FastAPI
fastapi==0.109.0
uvicorn==0.27.0

# MongoDB (async)
motor==3.3.2
pymongo==4.6.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation & Utils
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI Integration
openai==1.12.0

# HTTP Client
httpx==0.26.0
